finished FIRST DRAFT of refactoring file tree for ExpressCheckout.
