+++ src/containers/checkout/assets/utils/index.js
@@ -17,7 +17,7 @@ export squarePaymentForm from '../../../../services/utils/squarePaymentForm';
 export cleanOffTypename from '../../../../services/utils/cleanOffTypename';
 export checkForToast from '../../../../services/utils/checkForToast';
 export generateFinalForm from '../../../../services/utils/orderForm/generateFinalForm';
-export countryConstants from './countryConstants';
-export prefectureConstants from './prefectureConstants';
-export stateConstants from './stateConstants';
+export Countries from './countryConstants';
+export Prefectures from './prefectureConstants';
+export States from './stateConstants';
 export SqrPaymentForm from '../../../../services/utils/squarePaymentForm';
