--- src/containers/checkout/components/submitOrder/utilities.imports.js
@@ -1,2 +1,0 @@
-export squarePaymentForm from '../../../../services/utils/squarePaymentForm';
-export squareAPI from '../../../../services/api/square/';
